Refactored FastAPI backend
